// packages/sdk/src/module/sync/config/configKeys.ts
/*
 * @Author: kasni.huang (kasni.huang@ringcentral.com)
 * @Date: 2019-03-12 10:33:13
 * Copyright © RingCentral. All rights reserved.
 */
enum SYNC_CONFIG_KEYS {
  SOCKET_SERVER_HOST = 'socket_server_host',
  LAST_INDEX_TIMESTAMP = 'LAST_INDEX_TIMESTAMP',
  CAN_UPDATE_INDEX_TIME_STAMP = 'CAN_UPDATE_INDEX_TIME_STAMP',
  FETCHED_REMAINING = 'FETCHED_REMAINING',
  INDEX_SUCCEED = 'INDEX_SUCCEED',
  INDEX_START_LOCAL_TIME = 'INDEX_START_LOCAL_TIME',
  SOCKET_CONNECTED_LOCAL_TIME = 'SOCKET_CONNECTED_LOCAL_TIME',
}

const CONFIG_KEYS = {
  SOCKET_SERVER_HOST: 'SOCKET_SERVER_HOST',
  STATIC_HTTP_SERVER: 'STATIC_HTTP_SERVER',
};

export { SYNC_CONFIG_KEYS, CONFIG_KEYS };